Make /rewind command always available
