+++ docs/changelog/v3/3.0.0.md
@@ -2,6 +2,9 @@
 
 ## 3.0.0b7
 
+- **Fix: `/rewind` was invisible by default, and typing it sent "/rewind" to the model** (`src/zrb/llm/ui/base/conversation_commands.py`, `src/zrb/llm/ui/base/commands.py`, `src/zrb/llm/ui/default/ui.py`, `test/llm/ui/base/test_commands_rewind.py` new, `conftest.py` new, `test/llm/ui/test_ui_command_handlers.py`): rewind availability was expressed in three places — the handler returned `False` without a snapshot manager, `default/ui.py` blanked `rewind_commands` before building the completer, and the help panel wrapped its row in `if base_ui.snapshot_manager is not None`. Since `LLM_ENABLE_REWIND` defaults to off, the shipped default hid the command from both `/help` and tab-completion, and the handler's `False` let `dispatch_command` forward the literal text to the LLM as a chat message — so the feature was undiscoverable and misbehaved when found. It now follows `/voice`: one gate, in the handler, after the token match, printing a warning and consuming the input. The message names `{ENV_PREFIX}_LLM_ENABLE_REWIND` when that knob is the reason and points at the session/`LLM_SNAPSHOT_DIR` requirement when it is not. `default/ui.py` no longer knows any command name, and `dataclasses.replace` is gone from it.
+- **Tests: the `llm/ui/base` command tests share one `MockUI`** (`test/llm/ui/base/conftest.py` new, `test_commands_basics.py`, `test_commands_shell.py`, `test_commands_voice.py`): three byte-identical 95-line copies of the `BaseUI` stand-in had accumulated. `test_get_help_text_empty_commands` was also passing for the wrong reason — it cleared thirteen command lists but not `rewind_commands`, and only stayed green because the help row used to be gated on `snapshot_manager`.
+
 - **Refactor: the terminal UI stopped threading 16 slash-command parameters through three layers** (`src/zrb/llm/ui/default/ui.py`, `src/zrb/llm/ui/default/app/layout.py`, `src/zrb/llm/ui/default/app/completion/completer.py`, `test/llm/ui/default/app/test_layout.py`, `test/llm/ui/default/app/completion/test_completer_arguments.py`, `test_completer_commands.py`): every command's alias list was a separate named parameter passed `UI` → `create_input_field` → `InputCompleter` — `layout.py` declared 16 of them and immediately forwarded all 16 without reading any, and the two model-visibility toggles rode along the same way. All three signatures now take the `UIConfig` that already holds those fields (`BaseUI.<x>_commands` is a property reading and writing `_ui_config`, so the values are identical), leaving `layout.py` a one-line forward. `/rewind` keeps its one irregularity — it is only completed when snapshots are enabled — via `dataclasses.replace(self._ui_config, rewind_commands=[])`, a copy that leaves `self.rewind_commands` intact for the command handler itself. Adding a slash command no longer touches `layout.py` or `default/ui.py` at all: 8 files down to 5.
 
 - **Fix: overriding a slash-command handler on a custom UI did nothing** (`src/zrb/llm/ui/base/commands.py`, `src/zrb/llm/ui/base/ui.py`, `test/architecture/test_command_dispatch_surface.py` new, `docs/llm/llm-custom-ui.md`): `BaseUI` publishes a `handle_<name>_command` method per slash command and is the documented Level-3 extension point, but `_command_table()` bound each handler to the sub-part that owns the code (`self._conversation.handle_save_command`) rather than to the UI instance. That resolves the method once, at table-build time, off an object the subclass is not in the MRO of — so a subclass overriding `handle_save_command` saw its code run when it called the method itself and silently skipped when a user typed `/save`. All 15 table entries now bind `base_ui.handle_<name>_command`, including the `handle_custom_command` fallback and `handle_toggle_voice` (which was bound to the dispatcher). `_command_table` is now public `command_table`, reachable via a new `BaseUI.commands` property so a subclass can reorder or extend it; `BaseUICommands` gained `conversation`/`models`/`exec` properties for the same reason. With dispatch going through the facade, `BaseUICommands`' 24 pass-through methods had one caller each and are deleted (608 → 551 lines); `BaseUI`'s delegators now forward straight to the owning part. Guarded by a new architecture ratchet asserting every `command_table()` handler satisfies `handler.__self__ is ui`, plus a behavioural test that a subclass override actually runs on `dispatch_command`.
